Add files via upload
The files in the original commit 1) were not linted (yikes!), 2) contained a .py file that was a pasted together version of two stand alone files and had some redundant steps because of this and 3) did not focus on Python basics. This commit is associated with the second time I am giving this lecture, and I decided I needed to add more about getting up and running with Python.